+++ src/gpu/VirtualFluids_GPU/Kernel/Kernels/BasicKernels/AdvectionDiffusion/Incompressible/Mod27/ADIncomp27/ADIncomp27.cu
@@ -10,7 +10,7 @@ std::shared_ptr<ADIncomp27> ADIncomp27::getNewInstance(std::shared_ptr<Parameter
 
 void ADIncomp27::run()
 {
-	int size_Mat = para->getParD(level)->numberOfNodes;
+	int size_Mat = (int)para->getParD(level)->numberOfNodes;
 	int numberOfThreads = para->getParD(level)->numberofthreads;
 
 	int Grid = (size_Mat / numberOfThreads) + 1;
@@ -28,15 +28,16 @@ void ADIncomp27::run()
 	dim3 grid(Grid1, Grid2);
 	dim3 threads(numberOfThreads, 1, 1);
 
-	LB_Kernel_AD_Incomp_27 << < grid, threads >> >(	para->getParD(level)->diffusivity, 
-													para->getParD(level)->typeOfGridNode,
-													para->getParD(level)->neighborX, 
-													para->getParD(level)->neighborY, 
-													para->getParD(level)->neighborZ,
-													para->getParD(level)->distributions.f[0], 
-													para->getParD(level)->distributionsAD27.f[0], 
-													para->getParD(level)->numberOfNodes,
-													para->getParD(level)->isEvenTimestep);
+	LB_Kernel_AD_Incomp_27 <<< grid, threads >>>(
+		para->getParD(level)->diffusivity, 
+		para->getParD(level)->typeOfGridNode,
+		para->getParD(level)->neighborX, 
+		para->getParD(level)->neighborY, 
+		para->getParD(level)->neighborZ,
+		para->getParD(level)->distributions.f[0], 
+		para->getParD(level)->distributionsAD27.f[0], 
+		para->getParD(level)->numberOfNodes,
+		para->getParD(level)->isEvenTimestep);
 	getLastCudaError("LB_Kernel_AD_Incomp_27 execution failed");
 }
