+++ src/gpu/VirtualFluids_GPU/Kernel/Kernels/WaleKernels/FluidFlow/Compressible/CumulantK17Debug/WaleCumulantK17DebugComp.cu
@@ -10,7 +10,7 @@ std::shared_ptr<WaleCumulantK17DebugComp> WaleCumulantK17DebugComp::getNewInstan
 
 void WaleCumulantK17DebugComp::run()
 {
-	int size_Mat = para->getParD(level)->numberOfNodes;
+	int size_Mat = (int)para->getParD(level)->numberOfNodes;
 	int numberOfThreads = para->getParD(level)->numberofthreads;
 
 	//int Grid = size_Array / numberOfThreads;
@@ -32,34 +32,34 @@ void WaleCumulantK17DebugComp::run()
 	dim3 grid(Grid1, Grid2, 1);
 	dim3 threads(numberOfThreads, 1, 1);
 
-	LB_Kernel_WaleCumulantK17DebugComp << < grid, threads >> >(
-																		para->getParD(level)->omega,
-																		para->getParD(level)->typeOfGridNode,
-																		para->getParD(level)->neighborX,
-																		para->getParD(level)->neighborY,
-																		para->getParD(level)->neighborZ,
-																		para->getParD(level)->neighborInverse,
-																		para->getParD(level)->velocityX,
-																		para->getParD(level)->velocityY,
-																		para->getParD(level)->velocityZ,
-																		para->getParD(level)->distributions.f[0],
-																		para->getParD(level)->turbViscosity,
-																		para->getParD(level)->gSij,
-																		para->getParD(level)->gSDij,
-																		para->getParD(level)->gDxvx,
-																		para->getParD(level)->gDyvx,
-																		para->getParD(level)->gDzvx,
-																		para->getParD(level)->gDxvy,
-																		para->getParD(level)->gDyvy,
-																		para->getParD(level)->gDzvy,
-																		para->getParD(level)->gDxvz,
-																		para->getParD(level)->gDyvz,
-																		para->getParD(level)->gDzvz,
-																		para->getParD(level)->numberOfNodes,
-																		level,
-																		para->getForcesDev(),
-                                                                        para->getQuadricLimitersDev(),
-																		para->getParD(level)->isEvenTimestep);
+	LB_Kernel_WaleCumulantK17DebugComp <<< grid, threads >>>(
+		para->getParD(level)->omega,
+		para->getParD(level)->typeOfGridNode,
+		para->getParD(level)->neighborX,
+		para->getParD(level)->neighborY,
+		para->getParD(level)->neighborZ,
+		para->getParD(level)->neighborInverse,
+		para->getParD(level)->velocityX,
+		para->getParD(level)->velocityY,
+		para->getParD(level)->velocityZ,
+		para->getParD(level)->distributions.f[0],
+		para->getParD(level)->turbViscosity,
+		para->getParD(level)->gSij,
+		para->getParD(level)->gSDij,
+		para->getParD(level)->gDxvx,
+		para->getParD(level)->gDyvx,
+		para->getParD(level)->gDzvx,
+		para->getParD(level)->gDxvy,
+		para->getParD(level)->gDyvy,
+		para->getParD(level)->gDzvy,
+		para->getParD(level)->gDxvz,
+		para->getParD(level)->gDyvz,
+		para->getParD(level)->gDzvz,
+		para->getParD(level)->numberOfNodes,
+		level,
+		para->getForcesDev(),
+        para->getQuadricLimitersDev(),
+		para->getParD(level)->isEvenTimestep);
 	getLastCudaError("LB_Kernel_WaleCumulantK17DebugComp execution failed");
 }
